S4-Generative-ai-tools-models-architectures/4-retrieval-augmented-generation/chatbot_RAG_Langchain_Openai/chatbot_rag_project/requirements.txt
langchain>=0.0.300
openai>=0.27.0
python-dotenv
numpy
tqdm
